+++ LEGO1/mxvariabletable.cpp
@@ -31,8 +31,13 @@ void MxVariableTable::SetVariable(const char *p_key, const char *p_value)
     cursor.GetMatch(&var);
     var->SetValue(p_value);
   } else {
-    Add(var);
+    MxHashTable<MxVariable>::Add(var);
   }
+
+  // create new var object here?
+
+  // Resize inlined
+  // Add not
 }
 
 // OFFSET: LEGO1 0x100b7740
@@ -44,7 +49,12 @@ void MxVariableTable::SetVariable(MxVariable *var)
   if (found)
     cursor.DeleteMatch();
 
-  Add(var);
+  MxHashTable<MxVariable>::Add(var);
+
+  // we already have the new variable to add.
+
+  // Resize not
+  // Add inlined
 }
 
 // OFFSET: LEGO1 0x100b78f0
